requirements.txt
altair==4.2.2
attrs==23.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
entrypoints==0.4
et-xmlfile==1.1.0
Faker==26.3.0
favicon==0.7.0
Flask==3.0.3
fitness-tools==0.1.1
Flask-SocketIO==5.3.6
fonttools==4.53.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
htbuilder==0.6.2
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mdurl==0.1.2
more-itertools==10.4.0
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
patsy==0.5.6
pillow==10.3.0
plotly==5.21.0
plotly-express==0.4.1
prometheus_client==0.20.0
protobuf==4.25.3
pyarrow==16.0.0
pycparser==2.22
pydeck==0.9.0b1
Pygments==2.17.2
pymdown-extensions==10.9
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.0
requests==2.31.0
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.0
scipy==1.13.0
selenium==4.20.0
setuptools==69.5.1
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
st-annotated-text==4.0.1
st-theme==1.2.3
statsmodels==0.14.2
streamlit==1.37.1
streamlit-awesome-table==0.1.0
streamlit-aggrid==1.0.5
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.6
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
validators==0.33.0
watchdog==4.0.0
Werkzeug==3.0.2
wsproto==1.2.0
zipp==3.18.1
